feat: use ndv
