--- src/micromanager_gui/_widgets/_stack_viewer/__init__.py
@@ -1,4 +1,0 @@
-from ._mda_viewer import MDAViewer
-from ._stack_viewer import StackViewer
-
-__all__ = ["StackViewer", "MDAViewer"]
