+++ src/micromanager_gui/_widgets/_viewers/_mda_viewer/__init__.py
@@ -1,0 +1,3 @@
+from ._mda_viewer import MDAViewer
+
+__all__ = ["MDAViewer"]
