+++ src/micromanager_gui/_widgets/_viewers/_preview_viewer/__init__.py
@@ -1,0 +1,3 @@
+from ._preview_viewer import Preview
+
+__all__ = ["Preview"]
