--- test.ome.zarr/.zgroup
@@ -1,3 +1,0 @@
-{
-    "zarr_format": 2
-}
